Adding action menu inside the game.
in the action menu, player had been allow to choose explore and walk, when player choose to explore, they could explore the knowledge that hide in room they are current in, if they enter walk, the menu will jump to the direction menu and allow user to choose the direction they want to go. The map limitation had been added. Player will not allow to move out of the map.